refactor: Result에 있던 string관련 로직 OutputView로 이동
